extract interface for FuzzyScore
- extract interface for FuzzyScore to have possibility of adding new implementation
