+++ app/src/main/java/fr/neamar/kiss/CustomIconDialog.java
@@ -46,7 +46,8 @@ import fr.neamar.kiss.icons.IconPackXML;
 import fr.neamar.kiss.icons.SystemIconPack;
 import fr.neamar.kiss.normalizer.StringNormalizer;
 import fr.neamar.kiss.utils.DrawableUtils;
-import fr.neamar.kiss.utils.FuzzyScore;
+import fr.neamar.kiss.utils.fuzzy.FuzzyFactory;
+import fr.neamar.kiss.utils.fuzzy.FuzzyScore;
 import fr.neamar.kiss.utils.UserHandle;
 import fr.neamar.kiss.utils.Utilities;
 
@@ -311,7 +312,7 @@ public class CustomIconDialog extends DialogFragment {
             Collection<IconPackXML.DrawableInfo> drawables = ((IconPackXML) iconPack).getDrawableList();
             if (drawables != null) {
                 StringNormalizer.Result normalized = StringNormalizer.normalizeWithResult(mSearch.getText(), true);
-                FuzzyScore fuzzyScore = new FuzzyScore(normalized.codePoints);
+                FuzzyScore fuzzyScore = FuzzyFactory.createFuzzyScore(getActivity(), normalized.codePoints);
                 for (IconPackXML.DrawableInfo info : drawables) {
                     if (fuzzyScore.match(info.getDrawableName()).match)
                         mIconData.add(new IconData((IconPackXML) iconPack, info));
