+++ app/src/main/java/fr/neamar/kiss/result/PhoneResult.java
@@ -21,7 +21,7 @@ import fr.neamar.kiss.R;
 import fr.neamar.kiss.adapter.RecordAdapter;
 import fr.neamar.kiss.pojo.PhonePojo;
 import fr.neamar.kiss.ui.ListPopup;
-import fr.neamar.kiss.utils.FuzzyScore;
+import fr.neamar.kiss.utils.fuzzy.FuzzyScore;
 
 public class PhoneResult extends CallResult<PhonePojo> {
